skipflow in help screen caused a problem
helpscreen shows when a notification arrives
but instantly the askAccepted becomes true then skipFlow is activated.
so while awaiting sender's info, the screen moves to askMain flow.
So this caused a warning: react state cannot be updated when unmounted.

So removed the skipFlow in helpscreen since the helpscreen is not
detached from askFlow. it is separate screen. so it does not appear when
going back.

now screen goes to the helpscreen when notification arrives.