used willFocus event in helpscreen to update the notification
but when the screen is in helpscreen already, the nodification data is
not updated. no way to refresh the screen? it seems there is.

in debug node, notification works welll, but in release of ios, it is
not.
 - no notification when the app is foreground
 - no navigatino to helpscreen when app is backgroud
- app crashes when the app is closed

it is like the android case, too. when release mode the navigation did
not work.